rework customlogger with .level method
