+++ src/main/java/fr/uge/chargepointconfiguration/logs/technical/TechnicalLogDto.java
@@ -1,6 +1,6 @@
 package fr.uge.chargepointconfiguration.logs.technical;
 
-import fr.uge.chargepointconfiguration.logs.sealed.TechnicalLog;
+import fr.uge.chargepointconfiguration.logs.sealed.TechnicalLogEntity;
 import java.sql.Timestamp;
 import org.apache.logging.log4j.Level;
 
@@ -9,14 +9,14 @@ import org.apache.logging.log4j.Level;
  *
  * @param id Database id of the log stored.
  * @param date Date when the log had been created.
- * @param component {@link TechnicalLog.Component}
+ * @param component {@link TechnicalLogEntity.Component}
  * @param level {@link Level}
  * @param completeLog All the log in one String.
  */
 public record TechnicalLogDto(
     int id,
     Timestamp date,
-    TechnicalLog.Component component,
+    TechnicalLogEntity.Component component,
     String level,
     String completeLog) {
 }
