+++ src/main/java/fr/uge/chargepointconfiguration/logs/technical/TechnicalLogRepository.java
@@ -1,6 +1,6 @@
 package fr.uge.chargepointconfiguration.logs.technical;
 
-import fr.uge.chargepointconfiguration.logs.sealed.TechnicalLog;
+import fr.uge.chargepointconfiguration.logs.sealed.TechnicalLogEntity;
 import java.util.List;
 import org.apache.logging.log4j.Level;
 import org.springframework.data.domain.Page;
@@ -13,17 +13,17 @@ import org.springframework.stereotype.Repository;
  * Repository for technical log.
  */
 @Repository
-public interface TechnicalLogRepository extends CrudRepository<TechnicalLog, Integer>,
-      PagingAndSortingRepository<TechnicalLog, Integer> {
+public interface TechnicalLogRepository extends CrudRepository<TechnicalLogEntity, Integer>,
+      PagingAndSortingRepository<TechnicalLogEntity, Integer> {
 
   /**
    * Method to return all technical logs.
    *
    * @return a list of all technical logs.
    */
-  List<TechnicalLog> findAll();
+  List<TechnicalLogEntity> findAll();
 
-  Page<TechnicalLog> findAll(Pageable pageable);
+  Page<TechnicalLogEntity> findAll(Pageable pageable);
 
   /**
    * Method to return all technical logs by the component.
@@ -32,8 +32,8 @@ public interface TechnicalLogRepository extends CrudRepository<TechnicalLog, Int
    * @param level {@link Level}
    * @return the list of technical logs by component.
    */
-  List<TechnicalLog> findAllByComponentAndLevel(
-          TechnicalLog.Component component,
+  List<TechnicalLogEntity> findAllByComponentAndLevel(
+          TechnicalLogEntity.Component component,
           String level);
 
   /**
@@ -42,7 +42,7 @@ public interface TechnicalLogRepository extends CrudRepository<TechnicalLog, Int
    * @param level {@link Level}
    * @return the list of technical logs by criticality.
    */
-  List<TechnicalLog> findAllByLevel(String level);
+  List<TechnicalLogEntity> findAllByLevel(String level);
 
 
 }
