+++ src/main/java/fr/uge/chargepointconfiguration/logs/technical/TechnicalLogService.java
@@ -1,6 +1,6 @@
 package fr.uge.chargepointconfiguration.logs.technical;
 
-import fr.uge.chargepointconfiguration.logs.sealed.TechnicalLog;
+import fr.uge.chargepointconfiguration.logs.sealed.TechnicalLogEntity;
 import java.util.List;
 import org.apache.logging.log4j.Level;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,8 +20,8 @@ public class TechnicalLogService {
     this.technicalLogRepository = technicalLogRepository;
   }
 
-  public List<TechnicalLog>  getTechnicalLogByComponentAndLevel(
-      TechnicalLog.Component component,
+  public List<TechnicalLogEntity>  getTechnicalLogByComponentAndLevel(
+      TechnicalLogEntity.Component component,
       Level level) {
     return technicalLogRepository.findAllByComponentAndLevel(component, level.name());
   }
@@ -31,12 +31,12 @@ public class TechnicalLogService {
   }
 
   /**
-   * Search for {@link TechnicalLog} with a pagination.
+   * Search for {@link TechnicalLogEntity} with a pagination.
    *
    * @param pageable         The page requested
-   * @return the list of corresponding {@link TechnicalLog}
+   * @return the list of corresponding {@link TechnicalLogEntity}
    */
-  public List<TechnicalLog> getPage(PageRequest pageable) {
+  public List<TechnicalLogEntity> getPage(PageRequest pageable) {
     return technicalLogRepository.findAll(pageable)
           .stream().toList();
   }
